refactor: split authoptions and getUser into two files
