+++ src/actions/ombul/read.ts
@@ -1,7 +1,7 @@
 'use server'
 import { createActionError, createPrismaActionError } from '@/actions/error'
 import prisma from '@/prisma'
-import { getUser } from '@/auth'
+import { getUser } from '@/auth/user'
 import type { ActionReturn } from '@/actions/Types'
 import type { ExpandedOmbul } from './Types'
 import type { Ombul } from '@prisma/client'
